requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.2.0
bip_utils==2.9.2
bitarray==2.9.2
cachetools==4.2.4
cbor2==5.6.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coincurve==19.0.1
construct==2.10.68
construct-typing==0.5.6
crcmod==1.7
cytoolz==0.12.3
ecdsa==0.18.0
ed25519-blake2b==1.4.1
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
exceptiongroup==1.2.0
frozenlist==1.4.1
h11==0.14.0
hexbytes==0.3.1
httpcore==0.16.3
httpx==0.23.3
idna==3.6
jsonalias==0.1.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
lxml==5.2.1
mnemonic==0.21
multidict==6.0.5
parsimonious==0.9.0
pillow==10.3.0
protobuf==5.26.1
py-sr25519-bindings==0.2.0
pycparser==2.21
pycryptodome==3.20.0
PyNaCl==1.5.0
python-docx==1.1.0
pyunormalize==15.1.0
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rlp==4.0.0
rpds-py==0.18.0
six==1.16.0
sniffio==1.3.1
solana==0.32.0
solders==0.20.0
toolz==0.12.1
types-cachetools==4.2.10
typing_extensions==4.10.0
urllib3==2.2.1
web3==6.16.0
websockets==11.0.3
yarl==1.9.4
